chore: move witness-related logic to a separate module
